Streamline config classes: merge extended properties into existing LegendConfig and TooltipConfig, remove duplicate ChartOptionsExtended.cs
